side-effect-tracker/agent_app/pyproject.toml
[project]
name = "agent-app"
version = "0.1.0"
description = "Medical side effect tracker — ClinicalTrials.gov + optional LLM agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
]

[project.optional-dependencies]
agent = [
    "langchain[openai]>=1.2.9",
    "langgraph>=0.2.0",
]
arcade = [
    "langchain[openai]>=1.2.9",
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.2.1",
]
